Split agent_connection into multiple files
