+++ vcx/dummy-cloud-agent/src/actors/agent_connection/internal.rs
@@ -95,255 +95,12 @@ impl AgentConnection {
             .into_box()
     }
 
-    /// Used only on legacy V1 protocol_type
-    fn handle_create_message(&mut self,
-                             msg: CreateMessage,
-                             mut tail: Vec<A2AMessage>,
-                             sender_verkey: &str) -> ResponseActFuture<Self, Vec<A2AMessage>, Error> {
-        trace!("AgentConnection::handle_create_message >> {:?}, {:?}, {:?}",
-               msg, tail, sender_verkey);
-
-        let CreateMessage { mtype, send_msg, reply_to_msg_id, uid } = msg;
-
-        let sender_verkey = sender_verkey.to_string();
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(move |_, slf, _| {
-                match (mtype, tail.pop()) {
-                    (RemoteMessageType::ConnReq, Some(A2AMessage::Version1(A2AMessageV1::MessageDetail(MessageDetail::ConnectionRequest(detail))))) => {
-                        slf.handle_create_connection_request(detail, sender_verkey)
-                    }
-                    (RemoteMessageType::ConnReqAnswer, Some(A2AMessage::Version1(A2AMessageV1::MessageDetail(MessageDetail::ConnectionRequestAnswer(detail))))) => {
-                        slf.handle_create_connection_request_answer(detail,
-                                                                    reply_to_msg_id.clone(),
-                                                                    uid,
-                                                                    sender_verkey)
-                    }
-                    (RemoteMessageType::ConnReqRedirect, Some(A2AMessage::Version1(A2AMessageV1::MessageDetail(MessageDetail::ConnectionRequestRedirect(detail))))) => {
-                        slf.handle_create_connection_request_redirect(detail,
-                                                                      reply_to_msg_id.clone(),
-                                                                      uid,
-                                                                      sender_verkey)
-                    }
-                    (mtype @ _, Some(A2AMessage::Version1(A2AMessageV1::MessageDetail(MessageDetail::General(detail))))) => {
-                        slf.handle_create_general_message(mtype,
-                                                          detail,
-                                                          reply_to_msg_id.clone(),
-                                                          uid,
-                                                          sender_verkey)
-                    }
-                    _ => err_act!(slf, err_msg("Unsupported message."))
-                }
-                    .map(|(msg_uid, a2a_msgs), _, _| (msg_uid, a2a_msgs, reply_to_msg_id))
-            })
-            .and_then(move |(msg_uid, mut a2a_msgs, reply_to_msg_id), slf, _| {
-                slf.send_message_if_needed(send_msg, &msg_uid, reply_to_msg_id)
-                    .map(|mut sent_message| {
-                        a2a_msgs.append(&mut sent_message);
-                        a2a_msgs
-                    })
-                    .into_actor(slf)
-            })
-            .into_box()
-    }
-
-    /// NOTE: Used only on non-Aries didcomm
-    fn handle_connection_request_message(&mut self,
-                                         msg: ConnectionRequest,
-                                         sender_verkey: &str) -> ResponseActFuture<Self, Vec<A2AMessage>, Error> {
-        trace!("AgentConnection::handle_connection_request_message >> {:?}, {:?}", msg, sender_verkey);
-
-        let send_msg = msg.send_msg;
-        let reply_to_msg_id = msg.reply_to_msg_id.clone();
-        let msg_detail = msg.into();
-        let sender_verkey = sender_verkey.to_string();
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(move |_, slf, _| {
-                slf.handle_create_connection_request(msg_detail, sender_verkey)
-            })
-            .and_then(move |(msg_uid, a2a_msgs), slf, _| {
-                slf.send_message_if_needed(send_msg, &msg_uid, reply_to_msg_id)
-                    .map(|_| a2a_msgs)
-                    .into_actor(slf)
-            })
-            .into_box()
-    }
-
-    fn handle_create_connection_request(&mut self,
-                                        msg_detail: ConnectionRequestMessageDetail,
-                                        sender_verkey: String) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
-        trace!("AgentConnection::handle_create_connection_request >> {:?}, {:?}", msg_detail, sender_verkey);
-
-        ftry_act!(self, self.validate_connection_request(&msg_detail, &sender_verkey));
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(|_, slf, _|
-                slf.verify_agent_key_dlg_proof(&slf.user_pairwise_verkey, &msg_detail.key_dlg_proof)
-                    .map(|_| msg_detail)
-                    .into_actor(slf)
-            )
-            .map(|msg_detail, slf, _| {
-                slf.state.agent_key_dlg_proof = Some(msg_detail.key_dlg_proof.clone());
-
-                let sender_did = slf.user_pairwise_did.clone();
-                let msg = slf.create_and_store_internal_message(None,
-                                                                RemoteMessageType::ConnReq,
-                                                                MessageStatusCode::Created,
-                                                                &sender_did,
-                                                                None,
-                                                                None,
-                                                                Some(map! { "phone_no".to_string() => msg_detail.phone_no.clone() }),
-                                                                None,
-                                                                None);
-
-                (msg, msg_detail)
-            })
-            .map(move |(msg, msg_detail), slf, _| {
-                let messages = slf.build_invite_message(&msg, &msg_detail);
-                (msg.uid, messages)
-            })
-            .into_box()
-    }
-
-    /// NOTE: Used only on non-Aries didcomm
-    fn handle_connection_request_answer_message(&mut self,
-                                                msg: ConnectionRequestAnswer,
-                                                sender_verkey: &str) -> ResponseActFuture<Self, Vec<A2AMessage>, Error> {
-        trace!("AgentConnection::handle_connection_request_answer_message >> {:?}, {:?}", msg, sender_verkey);
-
-        let send_msg = msg.send_msg;
-        let reply_to_msg_id = msg.reply_to_msg_id.clone();
-        let msg_uid = msg.id.clone();
-        let sender_verkey = sender_verkey.to_string();
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(move |_, slf, _| {
-                slf.handle_create_connection_request_answer(msg.into(), reply_to_msg_id.clone(), Some(msg_uid), sender_verkey)
-                    .map(|(msg_uid, a2a_msgs), _, _| (msg_uid, a2a_msgs, reply_to_msg_id))
-            })
-            .and_then(move |(msg_uid, a2a_msgs, reply_to_msg_id), slf, _| {
-                slf.send_message_if_needed(send_msg, &msg_uid, reply_to_msg_id)
-                    .map(|_| a2a_msgs)
-                    .into_actor(slf)
-            })
-            .into_box()
-    }
-
-    /// NOTE: Used only on non-Aries didcomm
-    fn handle_create_connection_request_answer(&mut self,
-                                               msg_detail: ConnectionRequestAnswerMessageDetail,
-                                               reply_to_msg_id: Option<String>,
-                                               msg_uid: Option<String>,
-                                               sender_verkey: String) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
-        trace!("AgentConnection::handle_create_connection_request_answer >> {:?}, {:?}, {:?}, {:?}",
-               msg_detail, reply_to_msg_id, msg_uid, sender_verkey);
-
-        let reply_to_msg_id = ftry_act!(self, {
-            reply_to_msg_id.clone()
-                .ok_or(err_msg("Required field `reply_to_msg_id` is missed."))
-        });
-
-        ftry_act!(self, self.validate_connection_request_answer(&msg_detail, &reply_to_msg_id));
-
-        match self.get_message_handler_role(&sender_verkey) {
-            MessageHandlerRole::Owner =>
-                self.sender_handle_create_connection_request_answer(msg_detail, reply_to_msg_id),
-            MessageHandlerRole::Remote =>
-                self.receipt_handle_create_connection_request_answer(msg_detail, reply_to_msg_id, msg_uid)
-        }
-    }
-
-    fn handle_connection_request_redirect_message(&mut self,
-                                                  msg: ConnectionRequestRedirect,
-                                                  sender_verkey: &str) -> ResponseActFuture<Self, Vec<A2AMessage>, Error> {
-        trace!("AgentConnection::handle_connection_request_redirect_message >> {:?}, {:?}", msg, sender_verkey);
-
-        let send_msg = msg.send_msg;
-        let reply_to_msg_id = msg.reply_to_msg_id.clone();
-        let msg_uid = msg.id.clone();
-        let sender_verkey = sender_verkey.to_string();
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(move |_, slf, _| {
-                slf.handle_create_connection_request_redirect(msg.into(), reply_to_msg_id.clone(), Some(msg_uid), sender_verkey)
-                    .map(|(msg_uid, a2a_msgs), _, _| (msg_uid, a2a_msgs, reply_to_msg_id))
-            })
-            .and_then(move |(msg_uid, a2a_msgs, reply_to_msg_id), slf, _| {
-                slf.send_message_if_needed(send_msg, &msg_uid, reply_to_msg_id)
-                    .map(|_| a2a_msgs)
-                    .into_actor(slf)
-            })
-            .into_box()
-    }
-
-    fn handle_create_connection_request_redirect(&mut self,
-                                                 msg_detail: ConnectionRequestRedirectMessageDetail,
-                                                 reply_to_msg_id: Option<String>,
-                                                 msg_uid: Option<String>,
-                                                 sender_verkey: String) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
-        trace!("AgentConnection::handle_create_connection_request_redirect >> {:?}, {:?}, {:?}, {:?}",
-               msg_detail, reply_to_msg_id, msg_uid, sender_verkey);
-
-        let reply_to_msg_id = ftry_act!(self, {
-            reply_to_msg_id.clone()
-                .ok_or(err_msg("Required field `reply_to_msg_id` is missed."))
-        });
-
-        ftry_act!(self, self.validate_connection_request_redirect(&msg_detail, &reply_to_msg_id));
-
-        match self.get_message_handler_role(&sender_verkey) {
-            MessageHandlerRole::Owner =>
-                self.sender_handle_create_connection_request_redirect(msg_detail, reply_to_msg_id),
-            MessageHandlerRole::Remote =>
-                self.receipt_handle_create_connection_request_redirect(msg_detail, reply_to_msg_id, msg_uid)
-        }
-    }
-
-    fn handle_send_remote_message(&mut self,
-                                  msg: SendRemoteMessage,
-                                  sender_verkey: &str) -> ResponseActFuture<Self, Vec<A2AMessage>, Error> {
-        trace!("AgentConnection::handle_send_remote_message >> {:?}, {:?}", msg, sender_verkey);
-
-        let send_msg = msg.send_msg;
-        let mtype = msg.mtype.clone();
-        let uid = msg.id.clone();
-        let reply_to_msg_id = msg.reply_to_msg_id.clone();
-        let sender_verkey = sender_verkey.to_string();
-
-        let msg_ = ftry_act!(self, {serde_json::to_vec(&msg.msg)});
-
-        let msg_detail = GeneralMessageDetail {
-            msg: msg_,
-            title: msg.title,
-            detail: msg.detail,
-        };
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(move |_, slf, _| {
-                slf.handle_create_general_message(mtype, msg_detail, reply_to_msg_id.clone(), Some(uid), sender_verkey)
-                    .map(|(msg_uid, a2a_msgs), _, _| (msg_uid, a2a_msgs, reply_to_msg_id))
-            })
-            .and_then(move |(msg_uid, a2a_msgs, reply_to_msg_id), slf, _| {
-                slf.send_message_if_needed(send_msg, &msg_uid, reply_to_msg_id)
-                    .map(|_| a2a_msgs)
-                    .into_actor(slf)
-            })
-            .into_box()
-    }
-
-    fn handle_create_general_message(&mut self,
-                                     mtype: RemoteMessageType,
-                                     msg_detail: GeneralMessageDetail,
-                                     reply_to_msg_id: Option<String>,
-                                     uid: Option<String>,
-                                     sender_verkey: String) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
+    pub(super) fn handle_create_general_message(&mut self,
+                                                mtype: RemoteMessageType,
+                                                msg_detail: GeneralMessageDetail,
+                                                reply_to_msg_id: Option<String>,
+                                                uid: Option<String>,
+                                                sender_verkey: String) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
         trace!("AgentConnection::handle_create_general_message >> {:?}, {:?}, {:?}, {:?}, {:?}",
                mtype, msg_detail, reply_to_msg_id, uid, sender_verkey);
 
@@ -410,7 +167,7 @@ impl AgentConnection {
         ok_act!(self, vec![message])
     }
 
-    fn get_messages(&self, msg: GetMessages) -> Vec<GetMessagesDetailResponse> {
+    pub(super) fn get_messages(&self, msg: GetMessages) -> Vec<GetMessagesDetailResponse> {
         let GetMessages { exclude_payload, uids, status_codes } = msg;
 
         self.state.messages
@@ -528,316 +285,6 @@ impl AgentConnection {
             .map(|uids| (uids, status_code))
     }
 
-    /// NOTE: Used only on non-Aries didcomm
-    fn sender_handle_create_connection_request_answer(&mut self,
-                                                      msg_detail: ConnectionRequestAnswerMessageDetail,
-                                                      reply_to_msg_id: String) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
-        trace!("AgentConnection::initiator_handle_create_connection_request_answer >> {:?}, {:?}",
-               msg_detail, reply_to_msg_id);
-
-        let key_dlg_proof = ftry_act!(self, {
-            msg_detail.key_dlg_proof.clone()
-                .ok_or(err_msg("Required field `key_dlg_proof` is missed."))
-        });
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(|_, slf, _|
-                slf.verify_agent_key_dlg_proof(&slf.user_pairwise_verkey, &key_dlg_proof)
-                    .map(|_| (msg_detail, reply_to_msg_id, key_dlg_proof))
-                    .into_actor(slf)
-            )
-            .and_then(|(msg_detail, reply_to_msg_id, key_dlg_proof), slf, _|
-                slf.verify_agent_key_dlg_proof(&msg_detail.sender_detail.verkey, &msg_detail.sender_detail.agent_key_dlg_proof)
-                    .map(|_| (msg_detail, reply_to_msg_id, key_dlg_proof))
-                    .into_actor(slf)
-            )
-            .map(|(msg_detail, reply_to_msg_id, key_dlg_proof), slf, _| {
-                let conn_req_msg = slf.create_and_store_internal_message(
-                    Some(reply_to_msg_id.as_str()),
-                    RemoteMessageType::ConnReq,
-                    MessageStatusCode::Received,
-                    &msg_detail.sender_detail.did,
-                    None,
-                    None,
-                    None,
-                    None,
-                    None,
-                );
-
-                let sender_did = slf.user_pairwise_did.clone();
-                let answer_msg = slf.create_and_store_internal_message(
-                    None,
-                    RemoteMessageType::ConnReqAnswer,
-                    msg_detail.answer_status_code.clone(),
-                    &sender_did,
-                    Some(conn_req_msg.uid.as_str()),
-                    None,
-                    None,
-                    msg_detail.thread.clone(),
-                    None,
-                );
-                slf.state.agent_key_dlg_proof = Some(key_dlg_proof);
-
-                slf.state.remote_connection_detail = Some(RemoteConnectionDetail {
-                    forward_agent_detail: msg_detail.sender_agency_detail.clone(),
-                    agent_detail: AgentDetail {
-                        did: msg_detail.sender_detail.did.clone(),
-                        verkey: msg_detail.sender_detail.verkey.clone(),
-                    },
-                    agent_key_dlg_proof: msg_detail.sender_detail.agent_key_dlg_proof.clone(),
-                });
-
-                (msg_detail, reply_to_msg_id, answer_msg.uid)
-            })
-            .and_then(|(msg_detail, reply_to_msg_id, answer_msg_uid), slf, _| {
-                slf.answer_message(&reply_to_msg_id, &answer_msg_uid, &msg_detail.answer_status_code)
-                    .into_future()
-                    .map(|_| (msg_detail, answer_msg_uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_their_did(&msg_detail.sender_detail.did, &msg_detail.sender_detail.verkey)
-                    .map(|_| (msg_detail, uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_their_did(&msg_detail.sender_agency_detail.did, &msg_detail.sender_agency_detail.verkey)
-                    .map(|_| uid)
-                    .into_actor(slf)
-            })
-            .map(|uid, _, _| {
-                let message = match ProtocolType::get() {
-                    ProtocolTypes::V1 => A2AMessage::Version1(A2AMessageV1::MessageCreated(MessageCreated { uid: uid.clone() })),
-                    ProtocolTypes::V2 => A2AMessage::Version2(A2AMessageV2::ConnectionRequestAnswerResponse(ConnectionRequestAnswerResponse { id: uid.clone(), sent: true }))
-                };
-                (uid, vec![message])
-            })
-            .into_box()
-    }
-
-    /// Used only on non-Aries didcomm
-    fn receipt_handle_create_connection_request_answer(&mut self,
-                                                       msg_detail: ConnectionRequestAnswerMessageDetail,
-                                                       reply_to_msg_id: String,
-                                                       msg_uid: Option<String>) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
-        trace!("AgentConnection::receipt_handle_create_connection_request_answer >> {:?}, {:?}, {:?}",
-               msg_detail, reply_to_msg_id, msg_uid);
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(|_, slf, _|
-                slf.verify_agent_key_dlg_proof(&msg_detail.sender_detail.verkey, &msg_detail.sender_detail.agent_key_dlg_proof)
-                    .map(|_| (msg_detail, reply_to_msg_id))
-                    .into_actor(slf)
-            )
-            .map(move |(msg_detail, reply_to_msg_id), slf, _| {
-                let answer_msg = slf.create_and_store_internal_message(
-                    msg_uid.as_ref().map(String::as_str),
-                    RemoteMessageType::ConnReqAnswer,
-                    msg_detail.answer_status_code.clone(),
-                    &msg_detail.sender_detail.did,
-                    None,
-                    None,
-                    None,
-                    msg_detail.thread.clone(),
-                    None,
-                );
-
-                slf.state.remote_connection_detail = Some(RemoteConnectionDetail {
-                    forward_agent_detail: msg_detail.sender_agency_detail.clone(),
-                    agent_detail: AgentDetail {
-                        did: msg_detail.sender_detail.did.clone(),
-                        verkey: msg_detail.sender_detail.verkey.clone(),
-                    },
-                    agent_key_dlg_proof: msg_detail.sender_detail.agent_key_dlg_proof.clone(),
-                });
-
-                (msg_detail, reply_to_msg_id, answer_msg.uid)
-            })
-            .and_then(|(msg_detail, reply_to_msg_id, answer_msg_uid), slf, _| {
-                slf.answer_message(&reply_to_msg_id, &answer_msg_uid, &msg_detail.answer_status_code).into_future()
-                    .map(|_| (msg_detail, answer_msg_uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_their_did(&msg_detail.sender_detail.did, &msg_detail.sender_detail.verkey)
-                    .map(|_| (msg_detail, uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_their_did(&msg_detail.sender_agency_detail.did, &msg_detail.sender_agency_detail.verkey)
-                    .map(|_| (msg_detail, uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_payload_for_connection_request_answer(&uid, msg_detail)
-                    .map(|_, _, _| uid)
-            })
-            .map(|uid, _, _| {
-                let message = match ProtocolType::get() {
-                    ProtocolTypes::V1 => A2AMessage::Version1(A2AMessageV1::MessageCreated(MessageCreated { uid: uid.clone() })),
-                    ProtocolTypes::V2 => A2AMessage::Version2(A2AMessageV2::ConnectionRequestAnswerResponse(ConnectionRequestAnswerResponse { id: uid.clone(), sent: true }))
-                };
-                (uid, vec![message])
-            })
-            .into_box()
-    }
-
-    fn sender_handle_create_connection_request_redirect(&mut self,
-                                                        msg_detail: ConnectionRequestRedirectMessageDetail,
-                                                        reply_to_msg_id: String) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
-        trace!("AgentConnection::initiator_handle_create_connection_request_redirect >> {:?}, {:?}",
-               msg_detail, reply_to_msg_id);
-
-        let key_dlg_proof = ftry_act!(self, {
-            msg_detail.key_dlg_proof.clone()
-                .ok_or(err_msg("Required field `key_dlg_proof` is missed."))
-        });
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(|_, slf, _|
-                slf.verify_agent_key_dlg_proof(&slf.user_pairwise_verkey, &key_dlg_proof)
-                    .map(|_| (msg_detail, reply_to_msg_id, key_dlg_proof))
-                    .into_actor(slf)
-            )
-            .and_then(|(msg_detail, reply_to_msg_id, key_dlg_proof), slf, _|
-                slf.verify_agent_key_dlg_proof(&msg_detail.sender_detail.verkey, &msg_detail.sender_detail.agent_key_dlg_proof)
-                    .map(|_| (msg_detail, reply_to_msg_id, key_dlg_proof))
-                    .into_actor(slf)
-            )
-            .map(|(msg_detail, reply_to_msg_id, key_dlg_proof), slf, _| {
-                let conn_req_msg = slf.create_and_store_internal_message(
-                    Some(reply_to_msg_id.as_str()),
-                    RemoteMessageType::ConnReq,
-                    MessageStatusCode::Received,
-                    &msg_detail.sender_detail.did,
-                    None,
-                    None,
-                    None,
-                    None,
-                    None,
-                );
-
-                let sender_did = slf.user_pairwise_did.clone();
-                let answer_msg = slf.create_and_store_internal_message(
-                    None,
-                    RemoteMessageType::ConnReqRedirect,
-                    msg_detail.answer_status_code.clone(),
-                    &sender_did,
-                    Some(conn_req_msg.uid.as_str()),
-                    None,
-                    None,
-                    msg_detail.thread.clone(),
-                    Some(msg_detail.redirect_detail.clone()),
-                );
-                slf.state.agent_key_dlg_proof = Some(key_dlg_proof);
-
-                slf.state.remote_connection_detail = Some(RemoteConnectionDetail {
-                    forward_agent_detail: msg_detail.sender_agency_detail.clone(),
-                    agent_detail: AgentDetail {
-                        did: msg_detail.sender_detail.did.clone(),
-                        verkey: msg_detail.sender_detail.verkey.clone(),
-                    },
-                    agent_key_dlg_proof: msg_detail.sender_detail.agent_key_dlg_proof.clone(),
-                });
-
-                (msg_detail, reply_to_msg_id, answer_msg.uid)
-            })
-            .and_then(|(msg_detail, reply_to_msg_id, answer_msg_uid), slf, _| {
-                slf.answer_message(&reply_to_msg_id, &answer_msg_uid, &msg_detail.answer_status_code)
-                    .into_future()
-                    .map(|_| (msg_detail, answer_msg_uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_their_did(&msg_detail.sender_detail.did, &msg_detail.sender_detail.verkey)
-                    .map(|_| (msg_detail, uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_their_did(&msg_detail.sender_agency_detail.did, &msg_detail.sender_agency_detail.verkey)
-                    .map(|_| uid)
-                    .into_actor(slf)
-            })
-            .map(|uid, _, _| {
-                let message = match ProtocolType::get() {
-                    ProtocolTypes::V1 => A2AMessage::Version1(A2AMessageV1::MessageCreated(MessageCreated { uid: uid.clone() })),
-                    ProtocolTypes::V2 => A2AMessage::Version2(A2AMessageV2::ConnectionRequestRedirectResponse(ConnectionRequestRedirectResponse { id: uid.clone(), sent: true }))
-                };
-                (uid, vec![message])
-            })
-            .into_box()
-    }
-
-    fn receipt_handle_create_connection_request_redirect(&mut self,
-                                                         msg_detail: ConnectionRequestRedirectMessageDetail,
-                                                         reply_to_msg_id: String,
-                                                         msg_uid: Option<String>) -> ResponseActFuture<Self, (String, Vec<A2AMessage>), Error> {
-        trace!("AgentConnection::receipt_handle_create_connection_request_redirect >> {:?}, {:?}, {:?}",
-               msg_detail, reply_to_msg_id, msg_uid);
-
-        future::ok(())
-            .into_actor(self)
-            .and_then(|_, slf, _|
-                slf.verify_agent_key_dlg_proof(&msg_detail.sender_detail.verkey, &msg_detail.sender_detail.agent_key_dlg_proof)
-                    .map(|_| (msg_detail, reply_to_msg_id))
-                    .into_actor(slf)
-            )
-            .map(move |(msg_detail, reply_to_msg_id), slf, _| {
-                let answer_msg = slf.create_and_store_internal_message(
-                    msg_uid.as_ref().map(String::as_str),
-                    RemoteMessageType::ConnReqRedirect,
-                    msg_detail.answer_status_code.clone(),
-                    &msg_detail.sender_detail.did,
-                    None,
-                    None,
-                    None,
-                    msg_detail.thread.clone(),
-                    Some(msg_detail.redirect_detail.clone()),
-                );
-
-                slf.state.remote_connection_detail = Some(RemoteConnectionDetail {
-                    forward_agent_detail: msg_detail.sender_agency_detail.clone(),
-                    agent_detail: AgentDetail {
-                        did: msg_detail.sender_detail.did.clone(),
-                        verkey: msg_detail.sender_detail.verkey.clone(),
-                    },
-                    agent_key_dlg_proof: msg_detail.sender_detail.agent_key_dlg_proof.clone(),
-                });
-
-                (msg_detail, reply_to_msg_id, answer_msg.uid)
-            })
-            .and_then(|(msg_detail, reply_to_msg_id, answer_msg_uid), slf, _| {
-                slf.answer_message(&reply_to_msg_id, &answer_msg_uid, &msg_detail.answer_status_code).into_future()
-                    .map(|_| (msg_detail, answer_msg_uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_their_did(&msg_detail.sender_detail.did, &msg_detail.sender_detail.verkey)
-                    .map(|_| (msg_detail, uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_their_did(&msg_detail.sender_agency_detail.did, &msg_detail.sender_agency_detail.verkey)
-                    .map(|_| (msg_detail, uid))
-                    .into_actor(slf)
-            })
-            .and_then(|(msg_detail, uid), slf, _| {
-                slf.store_payload_for_connection_request_redirect(&uid, msg_detail)
-                    .map(|_, _, _| uid)
-            })
-            .map(|uid, _, _| {
-                let message = match ProtocolType::get() {
-                    ProtocolTypes::V1 => A2AMessage::Version1(A2AMessageV1::MessageCreated(MessageCreated { uid: uid.clone() })),
-                    ProtocolTypes::V2 => A2AMessage::Version2(A2AMessageV2::ConnectionRequestAnswerResponse(ConnectionRequestAnswerResponse { id: uid.clone(), sent: true }))
-                };
-                (uid, vec![message])
-            })
-            .into_box()
-    }
-
     /// Dispatches metadata about receive message in form of HTTP POST on specified url.
     /// Any HTTP error codes returned from the target URL are ignored. The HTTP request dispatch is
     /// non blocking.
@@ -877,35 +324,17 @@ impl AgentConnection {
         }
     }
 
-    /// The heart of Aries cross-domain communication. All incoming Aries messages are coming in
-    /// through this method.
-    fn handle_forward_message(&mut self, msg: ForwardV3) -> ResponseActFuture<Self, Vec<A2AMessage>, Error> {
-        let msg_ = ftry_act!(self, {serde_json::to_vec(&msg.msg)});
-
-        self.create_and_store_internal_message(None,
-                                               RemoteMessageType::Other(String::from("aries")),
-                                               MessageStatusCode::Received,
-                                               &String::new(),
-                                               None,
-                                               Some(msg_),
-                                               None,
-                                               None,
-                                               None);
-
-        ok_act!(self, vec![])
-    }
-
     /// Stores a message into agent's in-memory storage.
-    fn create_and_store_internal_message(&mut self,
-                                         uid: Option<&str>,
-                                         mtype: RemoteMessageType,
-                                         status_code: MessageStatusCode,
-                                         sender_did: &str,
-                                         ref_msg_id: Option<&str>,
-                                         payload: Option<Vec<u8>>,
-                                         sending_data: Option<HashMap<String, Option<String>>>,
-                                         thread: Option<Thread>,
-                                         redirect_detail: Option<RedirectDetail>) -> InternalMessage {
+    pub(super) fn create_and_store_internal_message(&mut self,
+                                                    uid: Option<&str>,
+                                                    mtype: RemoteMessageType,
+                                                    status_code: MessageStatusCode,
+                                                    sender_did: &str,
+                                                    ref_msg_id: Option<&str>,
+                                                    payload: Option<Vec<u8>>,
+                                                    sending_data: Option<HashMap<String, Option<String>>>,
+                                                    thread: Option<Thread>,
+                                                    redirect_detail: Option<RedirectDetail>) -> InternalMessage {
         trace!("AgentConnection::create_and_store_internal_message >> {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}",
                uid, mtype, status_code, sender_did, ref_msg_id, payload, sending_data, redirect_detail);
 
@@ -955,84 +384,22 @@ impl AgentConnection {
             .into_box()
     }
 
-    /// NOTE: Used only on non-Aries didcomm
-    fn store_payload_for_connection_request_answer(&mut self,
-                                                   msg_uid: &str,
-                                                   msg_detail: ConnectionRequestAnswerMessageDetail) -> ResponseActFuture<Self, (), Error> {
-        trace!("AgentConnection::store_payload_for_connection_request_answer >> {:?}, {:?}",
-               msg_uid, msg_detail);
-
-        if !self.state.messages.contains_key(msg_uid) {
-            return err_act!(self, err_msg("Message not found."));
-        }
-
-        let msg_uid = msg_uid.to_string();
-
-        self.build_payload_message(RemoteMessageType::ConnReqAnswer, &json!({"senderDetail": msg_detail.sender_detail}))
-            .into_actor(self)
-            .map(move |payload, slf, _|
-                slf.state.messages.get_mut(&msg_uid)
-                    .map(|message| message.payload = Some(payload))
-                    .unwrap()
-            )
-            .into_box()
-    }
-
-    fn store_payload_for_connection_request_redirect(&mut self,
-                                                     msg_uid: &str,
-                                                     msg_detail: ConnectionRequestRedirectMessageDetail) -> ResponseActFuture<Self, (), Error> {
-        trace!("AgentConnection::store_payload_for_connection_request_redirect >> {:?}, {:?}",
-               msg_uid, msg_detail);
-
-        if !self.state.messages.contains_key(msg_uid) {
-            return err_act!(self, err_msg("Message not found."));
-        }
-
-        let msg_uid = msg_uid.to_string();
-
-        // TODO: Darko: Fix
-        self.build_payload_message(RemoteMessageType::ConnReqRedirect, &json!({"redirectDetail": msg_detail.redirect_detail}))
-            .into_actor(self)
-            .map(move |payload, slf, _|
-                slf.state.messages.get_mut(&msg_uid)
-                    .map(|message| message.payload = Some(payload))
-                    .unwrap()
-            )
-            .into_box()
-    }
-
-    fn is_sent_by_owner(&self, sender_verkey: &str) -> bool {
+    pub(super) fn is_sent_by_owner(&self, sender_verkey: &str) -> bool {
         sender_verkey == self.user_pairwise_verkey
     }
 
-    fn is_sent_by_remote(&self, sender_verkey: &str) -> bool {
+    pub(super) fn is_sent_by_remote(&self, sender_verkey: &str) -> bool {
         match self.state.remote_connection_detail {
             Some(ref remote_connection_detail) => sender_verkey == remote_connection_detail.agent_key_dlg_proof.agent_delegated_key,
             None => true
         }
     }
 
-    fn get_message_handler_role(&self, sender_verkey: &str) -> MessageHandlerRole {
+    pub(super) fn get_message_handler_role(&self, sender_verkey: &str) -> MessageHandlerRole {
         trace!("AgentConnection::get_message_handler_role >> {:?}",
                sender_verkey);
 
         if self.is_sent_by_owner(sender_verkey) { MessageHandlerRole::Owner } else { MessageHandlerRole::Remote }
-    }
-
-    fn validate_connection_request(&self,
-                                   msg_detail: &ConnectionRequestMessageDetail,
-                                   sender_verkey: &str) -> Result<(), Error> {
-        trace!("AgentConnection::validate_connection_request >> {:?}, {:?}",
-               msg_detail, sender_verkey);
-
-        if !self.is_sent_by_owner(sender_verkey) {
-            return Err(err_msg("Unknown message sender."));
-        }
-
-        self.check_no_connection_established()?;
-        self.check_no_accepted_invitation_exists()?;
-
-        Ok(())
     }
 
     fn check_message_sender(&self, msg: Option<&A2AMessage>, sender_verkey: &str) -> Result<(), Error> {
@@ -1066,37 +433,7 @@ impl AgentConnection {
         Err(err_msg("Invalid message sender."))
     }
 
-    fn validate_connection_request_answer(&self,
-                                          msg_detail: &ConnectionRequestAnswerMessageDetail,
-                                          reply_to_msg_id: &str) -> Result<(), Error> {
-        trace!("AgentConnection::validate_connection_request_answer >> {:?}, {:?}",
-               msg_detail, reply_to_msg_id);
-
-        self.check_no_accepted_invitation_exists()?;
-        self.check_valid_status_code(&msg_detail.answer_status_code)?;
-
-        if let Some(msg) = self.state.messages.get(reply_to_msg_id) {
-            self.check_if_message_not_already_answered(&msg.status_code)?;
-        }
-        Ok(())
-    }
-
-    fn validate_connection_request_redirect(&self,
-                                            msg_detail: &ConnectionRequestRedirectMessageDetail,
-                                            reply_to_msg_id: &str) -> Result<(), Error> {
-        trace!("AgentConnection::validate_connection_request_redirect >> {:?}, {:?}",
-               msg_detail, reply_to_msg_id);
-
-        self.check_no_accepted_invitation_exists()?;
-        self.check_valid_status_code(&msg_detail.answer_status_code)?;
-
-        if let Some(msg) = self.state.messages.get(reply_to_msg_id) {
-            self.check_if_message_not_already_answered(&msg.status_code)?;
-        }
-        Ok(())
-    }
-
-    fn validate_general_message(&self, reply_to_msg_id: Option<&str>) -> Result<(), Error> {
+    pub(super) fn validate_general_message(&self, reply_to_msg_id: Option<&str>) -> Result<(), Error> {
         trace!("AgentConnection::validate_general_message >> {:?}", reply_to_msg_id);
 
         if let Some(msg_id) = reply_to_msg_id {
@@ -1116,7 +453,7 @@ impl AgentConnection {
             .map(|message| message.status_code = status.clone())
     }
 
-    fn answer_message(&mut self, uid: &str, ref_msg_id: &str, status_code: &MessageStatusCode) -> Result<(), Error> {
+    pub(super) fn answer_message(&mut self, uid: &str, ref_msg_id: &str, status_code: &MessageStatusCode) -> Result<(), Error> {
         trace!("AgentConnection::answer_message >> {:?}, {:?}, {:?}", uid, ref_msg_id, status_code);
 
         self.state.messages.get_mut(uid)
@@ -1127,7 +464,7 @@ impl AgentConnection {
             })
     }
 
-    fn store_their_did(&self, did: &str, verkey: &str) -> ResponseFuture<(), Error> {
+    pub(super) fn store_their_did(&self, did: &str, verkey: &str) -> ResponseFuture<(), Error> {
         trace!("AgentConnection::store_their_did >> {:?}, {:?}", did, verkey);
 
         let their_did_info = json!({
@@ -1144,7 +481,7 @@ impl AgentConnection {
             .into_box()
     }
 
-    fn check_if_message_not_already_answered(&self, status_code: &MessageStatusCode) -> Result<(), Error> {
+    pub(super) fn check_if_message_not_already_answered(&self, status_code: &MessageStatusCode) -> Result<(), Error> {
         trace!("AgentConnection::check_if_message_not_already_answered >> {:?}", status_code);
 
         if MessageStatusCode::valid_status_codes().contains(status_code) {
@@ -1153,7 +490,7 @@ impl AgentConnection {
         Ok(())
     }
 
-    fn check_valid_status_code(&self, status_code: &MessageStatusCode) -> Result<(), Error> {
+    pub(super) fn check_valid_status_code(&self, status_code: &MessageStatusCode) -> Result<(), Error> {
         trace!("AgentConnection::check_valid_status_code >> {:?}", status_code);
 
         if !MessageStatusCode::valid_status_codes().contains(status_code) {
@@ -1162,7 +499,7 @@ impl AgentConnection {
         Ok(())
     }
 
-    fn check_no_connection_established(&self) -> Result<(), Error> {
+    pub(super) fn check_no_connection_established(&self) -> Result<(), Error> {
         trace!("AgentConnection::check_no_connection_established >>");
 
         if self.state.remote_connection_detail.is_some() {
@@ -1171,7 +508,7 @@ impl AgentConnection {
         Ok(())
     }
 
-    fn check_no_accepted_invitation_exists(&self) -> Result<(), Error> {
+    pub(super) fn check_no_accepted_invitation_exists(&self) -> Result<(), Error> {
         trace!("AgentConnection::check_no_accepted_invitation_exists >>");
 
         let is_exists = self.state.messages.values()
@@ -1184,7 +521,7 @@ impl AgentConnection {
         Ok(())
     }
 
-    fn verify_agent_key_dlg_proof(&self, sender_verkey: &str, key_dlg_proof: &KeyDlgProof) -> ResponseFuture<(), Error> {
+    pub(super) fn verify_agent_key_dlg_proof(&self, sender_verkey: &str, key_dlg_proof: &KeyDlgProof) -> ResponseFuture<(), Error> {
         trace!("AgentConnection::verify_agent_key_dlg_proof >> {:?}, {:?}",
                sender_verkey, key_dlg_proof);
 
@@ -1217,7 +554,7 @@ impl AgentConnection {
         Ok(())
     }
 
-    fn send_message_if_needed(&mut self,
+    pub(super) fn send_message_if_needed(&mut self,
                               send_msg: bool,
                               uid: &str,
                               reply_to_msg_id: Option<String>) -> ResponseFuture<Vec<A2AMessage>, Error> {
@@ -1278,22 +615,6 @@ impl AgentConnection {
         self.send_remote_message(message, endpoint)
     }
 
-    fn send_invite_redirect_message(&mut self, message: InternalMessage, reply_to: Option<String>) -> ResponseFuture<(), Error> {
-        trace!("AgentConnection::send_invite_redirect_message >> {:?}, {:?}",
-               message, reply_to);
-
-        let reply_to = ftry!(reply_to.ok_or(err_msg("Missed required field `reply_to_msg_id`.")));
-
-        if message.status_code != MessageStatusCode::Redirected {
-            return err!(err_msg("Message status is not redirected."));
-        }
-
-        let invite_redirect = ftry!(self.build_invite_redirect_message(message, &reply_to));
-        let message = ftry!(self.prepare_remote_message(invite_redirect));
-        let endpoint = ftry!(self.get_remote_endpoint());
-        self.send_remote_message(message, endpoint)
-    }
-
     fn send_general_message(&self, message: InternalMessage, reply_to: Option<&str>) -> ResponseFuture<(), Error> {
         trace!("AgentConnection::send_general_message >> {:?}, {:?}",
                message, reply_to);
@@ -1310,7 +631,7 @@ impl AgentConnection {
         }
     }
 
-    fn send_remote_message(&self, message: Vec<u8>, endpoint: String) -> ResponseFuture<(), Error> {
+    pub(super) fn send_remote_message(&self, message: Vec<u8>, endpoint: String) -> ResponseFuture<(), Error> {
         let router = self.router.clone();
         future::ok(())
             .and_then(move |_| {
@@ -1322,13 +643,13 @@ impl AgentConnection {
             .into_box()
     }
 
-    fn get_remote_endpoint(&self) -> Result<String, Error> {
+    pub(super) fn get_remote_endpoint(&self) -> Result<String, Error> {
         let endpoint = self.state.remote_connection_detail.as_ref().map(|detail| detail.forward_agent_detail.endpoint.to_string())
             .ok_or(err_msg("Missed Remote Connection Details."))?;
         Ok(format!("{}/agency/msg", endpoint))
     }
 
-    fn prepare_remote_message(&self, message: Vec<A2AMessage>) -> Result<Vec<u8>, Error> {
+    pub(super) fn prepare_remote_message(&self, message: Vec<A2AMessage>) -> Result<Vec<u8>, Error> {
         trace!("AgentConnection::prepare_remote_message >> {:?}", message);
 
         let remote_connection_detail = self.state.remote_connection_detail.as_ref()
@@ -1348,7 +669,7 @@ impl AgentConnection {
         Ok(message)
     }
 
-    fn build_payload_message<T>(&self, type_: RemoteMessageType, msg: &T) -> ResponseFuture<Vec<u8>, Error> where T: ::serde::Serialize + ::std::fmt::Debug {
+    pub(super) fn build_payload_message<T>(&self, type_: RemoteMessageType, msg: &T) -> ResponseFuture<Vec<u8>, Error> where T: ::serde::Serialize + ::std::fmt::Debug {
         trace!("AgentConnection::build_payload_message >> {:?}, {:?}",
                type_, msg);
 
@@ -1387,7 +708,7 @@ impl AgentConnection {
         }
     }
 
-    fn build_forward_message(&self, fwd: &str, message: Vec<u8>) -> Result<Vec<A2AMessage>, Error> {
+    pub(super) fn build_forward_message(&self, fwd: &str, message: Vec<u8>) -> Result<Vec<A2AMessage>, Error> {
         trace!("AgentConnection::build_forward_message >> {:?}, {:?}",
                fwd, message);
 
@@ -1402,7 +723,7 @@ impl AgentConnection {
         Ok(vec![message])
     }
 
-    fn build_invite_message(&self, msg: &InternalMessage, msg_detail: &ConnectionRequestMessageDetail) -> Vec<A2AMessage> {
+    pub(super) fn build_invite_message(&self, msg: &InternalMessage, msg_detail: &ConnectionRequestMessageDetail) -> Vec<A2AMessage> {
         trace!("AgentConnection::build_invite_message >> {:?}, {:?}",
                msg, msg_detail);
 
@@ -1498,65 +819,6 @@ impl AgentConnection {
                         thread: message.thread.clone().unwrap_or(Thread::new()),
                     };
                     vec![A2AMessage::Version2(A2AMessageV2::ConnectionRequestAnswer(msg))]
-                }
-            };
-
-        Ok(messages)
-    }
-
-    fn build_invite_redirect_message(&self, message: InternalMessage, reply_to: &str) -> Result<Vec<A2AMessage>, Error> {
-        trace!("AgentConnection::build_invite_redirect_message >> {:?}, {:?}",
-               message, reply_to);
-
-        let agent_key_dlg_proof = self.state.agent_key_dlg_proof.clone()
-            .ok_or(err_msg("Missed Key Delegation Proof."))?;
-
-        let sender_detail = SenderDetail {
-            did: self.user_pairwise_did.clone(),
-            verkey: self.user_pairwise_verkey.clone(),
-            agent_key_dlg_proof,
-            name: None,
-            logo_url: None,
-            public_did: None,
-        };
-
-        let redirect_detail = message.redirect_detail.ok_or(err_msg("Missed Redirect Detail."))?;
-
-        let messages =
-            match ProtocolType::get() {
-                ProtocolTypes::V1 => {
-                    let msg_create = CreateMessage {
-                        mtype: message._type.clone(),
-                        send_msg: false,
-                        uid: Some(message.uid.clone()),
-                        reply_to_msg_id: Some(reply_to.to_string()),
-                    };
-
-                    let msg_detail = ConnectionRequestRedirectMessageDetail {
-                        key_dlg_proof: None,
-                        sender_detail,
-                        redirect_detail,
-                        sender_agency_detail: self.forward_agent_detail.clone(),
-                        answer_status_code: MessageStatusCode::Redirected,
-                        thread: None,
-                    };
-
-                    vec![A2AMessage::Version1(A2AMessageV1::CreateMessage(msg_create)),
-                         A2AMessage::Version1(A2AMessageV1::MessageDetail(MessageDetail::ConnectionRequestRedirect(msg_detail)))]
-                }
-                ProtocolTypes::V2 => {
-                    let msg = ConnectionRequestRedirect {
-                        send_msg: false,
-                        id: message.uid.clone(),
-                        reply_to_msg_id: Some(reply_to.to_string()),
-                        key_dlg_proof: None,
-                        sender_detail,
-                        redirect_detail,
-                        sender_agency_detail: self.forward_agent_detail.clone(),
-                        answer_status_code: MessageStatusCode::Redirected,
-                        thread: message.thread.clone().unwrap_or(Thread::new()),
-                    };
-                    vec![A2AMessage::Version2(A2AMessageV2::ConnectionRequestRedirect(msg))]
                 }
             };
