+++ vcx/dummy-cloud-agent/src/actors/agent_connection/mod.rs
@@ -1,2 +1,5 @@
 pub mod agent_connection;
 mod internal;
+mod internal_legacy;
+mod internal_aries;
+mod internal_redirect;
